Fixed sampling of colour pairs based on probability matrix
Previous script was improperly sampling what colour pairs would be
shown on each trial. Due to the condition that each trial required one
of each colour, the while sampling look made it most likely that all
high-probability pairs would be shown on a trial. This was fixed by
updating the probability matrix as colours were selected, ensuring the
probability of any pair that contained the already selected colours to
be zero.