Added awareness test
Added an objective awareness test to the end of the experiment, where
each colour is presented and the participant is asked to select which
colour was most likely to occur with this colour. The analysis code has
been updated to test awareness.

Additionally: A bug where participants pressed two buttons
simultaneously would crash the logical test has been fixed.